Move AMI functions to lib.
